:recycle: [#50] move json filter functions to separate files
