+++ src/open_producten/producten/viewsets/product.py
@@ -1,4 +1,3 @@
-from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 import django_filters
@@ -9,32 +8,20 @@ from open_producten.logging.api_tools import AuditTrailViewSetMixin
 from open_producten.producten.kanalen import KANAAL_PRODUCTEN
 from open_producten.producten.models import Product
 from open_producten.producten.serializers.product import ProductSerializer
+from open_producten.utils.enums import Operators
 from open_producten.utils.filters import (
     FilterSet,
     ManyCharFilter,
-    Operators,
     TranslationFilter,
     filter_data_attr_value_part,
-    validate_data_attr,
 )
+from open_producten.utils.helpers import display_choice_values_for_help_text
+from open_producten.utils.validators import validate_data_attr
 from open_producten.utils.views import OrderedModelViewSet
-
-
-def display_choice_values_for_help_text(Choices: type[models.TextChoices]) -> str:
-    items = []
-
-    for key, value in Choices.choices:
-        item = f"* `{key}` - {value}"
-        items.append(item)
-
-    return "\n".join(items)
-
 
 DATA_ATTR_HELP_TEXT = _(
     """
-Only include objects that have attributes with certain values.
-
-een json filter parameter heeft de format `key__operator__waarde`.
+Een json filter parameter heeft de format `key__operator__waarde`.
 `key` is de naam van de attribuut, `operator` is de operator die gebruikt moet worden en `waarde` is de waarde waarop zal worden gezocht.
 
 Waardes kunnen een string, nummer of datum (ISO format; YYYY-MM-DD) zijn.
