+++ src/open_producten/utils/filters.py
@@ -1,13 +1,11 @@
-from datetime import date
-
 from django import forms
-from django.core.validators import RegexValidator
 from django.db import models
-from django.utils.translation import gettext_lazy as _
 
 import django_filters
 from django_filters.rest_framework import DjangoFilterBackend, FilterSet as _FilterSet
-from rest_framework import serializers
+
+from open_producten.utils.enums import Operators
+from open_producten.utils.helpers import string_to_value
 
 
 class FilterSet(_FilterSet):
@@ -60,13 +58,6 @@ class ManyCharFilter(django_filters.CharFilter):
     field_class = ManyCharField
 
 
-class ManyRegexValidator(RegexValidator):
-
-    def __call__(self, values):
-        for value in values:
-            super().__call__(value)
-
-
 class TranslationFilter(django_filters.CharFilter):
     """
 
@@ -107,84 +98,6 @@ class TranslationFilter(django_filters.CharFilter):
         return qs
 
 
-class Operators(models.TextChoices):
-    exact = "exact", _("gelijk aan")
-    gt = "gt", _("grooter dan")
-    gte = "gte", _("grooter dan of gelijk aan")
-    lt = "lt", _("kleiner dan")
-    lte = "lte", _("kleiner dan of gelijk aan")
-    icontains = "icontains", _("case-insensitive partial match")
-    in_list = "in", _("in een lijst van waarden gescheiden door `|`")
-
-
-def validate_data_attr_value_part(value_part: str, code: str):
-    try:
-        variable, operator, val = value_part.rsplit("__", 2)
-    except ValueError:
-        message = _(
-            "Filter '%(value_part)s' heeft niet de format 'key__operator__waarde'"
-        ) % {"value_part": value_part}
-        raise serializers.ValidationError(message, code=code)
-
-    if operator not in Operators.values:
-        message = _("operator `%(operator)s` is niet bekend/ondersteund") % {
-            "operator": operator
-        }
-        raise serializers.ValidationError(message, code=code)
-
-    if operator not in (
-        Operators.exact,
-        Operators.icontains,
-        Operators.in_list,
-    ) and isinstance(string_to_value(val), str):
-        message = _(
-            "Operator `%(operator)s` ondersteund alleen datums and/or numerieke waarden"
-        ) % {"operator": operator}
-        raise serializers.ValidationError(message, code=code)
-
-
-def validate_data_attr(value: list):
-    code = "invalid-data-attr-query"
-
-    for value_part in value:
-        # check that comma can be only in the value part
-        if "," in value_part.rsplit("__", 1)[0]:
-            message = _(
-                "Filter '%(value_part)s' moet de format 'key__operator__value' hebben, "
-                "komma's kunnen alleen in de `waarde` worden toegevoegd"
-            ) % {"value_part": value_part}
-            raise serializers.ValidationError(message, code=code)
-
-        validate_data_attr_value_part(value_part, code)
-
-
-def string_to_value(value: str) -> str | float | date:
-    if is_number(value):
-        return float(value)
-    elif is_date(value):
-        return date.fromisoformat(value)
-
-    return value
-
-
-def is_date(value: str) -> bool:
-    try:
-        date.fromisoformat(value)
-    except ValueError:
-        return False
-
-    return True
-
-
-def is_number(value: str) -> bool:
-    try:
-        float(value)
-    except ValueError:
-        return False
-
-    return True
-
-
 def filter_data_attr_value_part(
     value_part: str, field_name: str, queryset: models.QuerySet
 ) -> models.QuerySet:
@@ -194,31 +107,32 @@ def filter_data_attr_value_part(
     variable, operator, str_value = value_part.rsplit("__", 2)
     real_value = string_to_value(str_value)
 
-    if operator == "exact":
-        #  for exact operator try to filter on string and numeric values
-        in_vals = [str_value]
-        if real_value != str_value:
-            in_vals.append(real_value)
-        queryset = queryset.filter(**{f"{field_name}__{variable}__in": in_vals})
-    elif operator == "icontains":
-        # icontains treats everything like strings
-        queryset = queryset.filter(
-            **{f"{field_name}__{variable}__icontains": str_value}
-        )
-    elif operator == "in":
-        # in must be a list
-        values = str_value.split("|")
-        queryset = queryset.filter(
-            **{
-                f"{field_name}__{variable}__in": [
-                    string_to_value(value) for value in values
-                ]
-            }
-        )
+    match operator:
+        case Operators.EXACT.value:
+            #  for exact operator try to filter on string and numeric values
+            in_vals = [str_value]
+            if real_value != str_value:
+                in_vals.append(real_value)
+            queryset = queryset.filter(**{f"{field_name}__{variable}__in": in_vals})
+        case Operators.ICONTAINS.value:
+            # icontains treats everything like strings
+            queryset = queryset.filter(
+                **{f"{field_name}__{variable}__icontains": str_value}
+            )
+        case Operators.IN_LIST.value:
+            # in must be a list
+            values = str_value.split("|")
+            queryset = queryset.filter(
+                **{
+                    f"{field_name}__{variable}__in": [
+                        string_to_value(value) for value in values
+                    ]
+                }
+            )
 
-    else:
-        # gt, gte, lt, lte operators
-        queryset = queryset.filter(
-            **{f"{field_name}__{variable}__{operator}": real_value}
-        )
+        case _:
+            # gt, gte, lt, lte operators
+            queryset = queryset.filter(
+                **{f"{field_name}__{variable}__{operator}": real_value}
+            )
     return queryset
