+++ src/open_producten/utils/validators.py
@@ -3,6 +3,10 @@ from django.core.validators import RegexValidator
 from django.utils.encoding import force_str
 from django.utils.translation import gettext_lazy as _
 
+from rest_framework import serializers
+
+from open_producten.utils.filters import Operators, string_to_value
+
 
 def validate_charfield_entry(value, allow_apostrophe=False):
     """
@@ -54,3 +58,51 @@ validate_postal_code = CustomRegexValidator(
         "Invalid postal code. A postal code must consist of 4 numbers followed by a space and two capital letters (e.g. 1234 AB)."
     ),
 )
+
+
+def validate_data_attr_value_part(value_part: str, code: str):
+    try:
+        variable, operator, val = value_part.rsplit("__", 2)
+    except ValueError:
+        message = _(
+            "Filter '%(value_part)s' heeft niet de format 'key__operator__waarde'"
+        ) % {"value_part": value_part}
+        raise serializers.ValidationError(message, code=code)
+
+    if operator not in Operators.values:
+        message = _("operator `%(operator)s` is niet bekend/ondersteund") % {
+            "operator": operator
+        }
+        raise serializers.ValidationError(message, code=code)
+
+    if operator not in (
+        Operators.EXACT,
+        Operators.ICONTAINS,
+        Operators.IN_LIST,
+    ) and isinstance(string_to_value(val), str):
+        message = _(
+            "Operator `%(operator)s` ondersteunt alleen datums en/of numerieke waarden"
+        ) % {"operator": operator}
+        raise serializers.ValidationError(message, code=code)
+
+
+def validate_data_attr(value: list):
+    code = "invalid-data-attr-query"
+
+    for value_part in value:
+        # check that comma can be only in the value part
+        if "," in value_part.rsplit("__", 1)[0]:
+            message = _(
+                "Filter '%(value_part)s' moet de format 'key__operator__value' hebben, "
+                "komma's kunnen alleen in de `waarde` worden toegevoegd"
+            ) % {"value_part": value_part}
+            raise serializers.ValidationError(message, code=code)
+
+        validate_data_attr_value_part(value_part, code)
+
+
+class ManyRegexValidator(RegexValidator):
+
+    def __call__(self, values):
+        for value in values:
+            super().__call__(value)
